Added tests to AuthorController class,  moved some code from AuthorController  to AuthorService class

+++ library/src/main/java/org/library/features/author/AuthorService.java
@@ -8,6 +8,7 @@ import java.util.List;
 public class AuthorService {
     private AuthorDAO authorDAO;
     private List<Author> authorList;
+    private String message;
 
     private void initializeAuthorDAO() {
         if (authorDAO == null) {
@@ -16,15 +17,32 @@ public class AuthorService {
         authorDAO.setSessionFactory(HibernateUtil.getSessionFactory());
     }
 
-    protected List<Author> getAuthorList(Login login){
+    protected List<Author> getAuthorList(Login login) {
         initializeAuthorDAO();
         authorList = authorDAO.getAuthorsList(login);
         return authorList;
     }
-    protected Author getAuthor(int id){
+
+    protected Author getAuthor(int id) {
         return authorList.stream().filter(author -> author.getId() == id).findAny().orElse(null);
     }
-    protected void delete(Author author){
-        authorDAO.delete(author);
+
+    protected boolean deleteIfPossible(int id) {
+        Author author = getAuthor(id);
+        if (author.getBooks().size() == 0) {
+            authorDAO.delete(author);
+            return true;
+        } else {
+            setMessage("Cannot delete selected author. There are some books assigned to this author, delete these books or change the author first.");
+            return false;
+        }
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
     }
 }
